requirements.txt
cloudinary>=1.44.1
django-cloudinary-storage>=0.3.0
django-bootstrap-v5
psycopg2-binary
django-allauth
channels
Pillow
daphne
channels-redis
redis
PyJWT
gunicorn
dj-database-url
python-dotenv
whitenoise
django-tinymce==4.1.0